Fix an import/export issue
